Revert byte pageShift to int, change method name doNotDestroy() to isDoNotDestroy(), change the field PoolArena.smallSubpagePools to package-private

+++ buffer/src/main/java/io/netty/buffer/PooledByteBufAllocator.java
@@ -295,7 +295,7 @@ public class PooledByteBufAllocator extends AbstractByteBufAllocator implements 
                     + directMemoryCacheAlignment + " (expected: power of two)");
         }
 
-        byte pageShifts = validateAndCalculatePageShifts(pageSize, directMemoryCacheAlignment);
+        int pageShifts = validateAndCalculatePageShifts(pageSize, directMemoryCacheAlignment);
 
         if (nHeapArena > 0) {
             heapArenas = newArenaArray(nHeapArena);
@@ -333,7 +333,7 @@ public class PooledByteBufAllocator extends AbstractByteBufAllocator implements 
         return new PoolArena[size];
     }
 
-    private static byte validateAndCalculatePageShifts(int pageSize, int alignment) {
+    private static int validateAndCalculatePageShifts(int pageSize, int alignment) {
         if (pageSize < MIN_PAGE_SIZE) {
             throw new IllegalArgumentException("pageSize: " + pageSize + " (expected: " + MIN_PAGE_SIZE + ')');
         }
@@ -348,7 +348,7 @@ public class PooledByteBufAllocator extends AbstractByteBufAllocator implements 
         }
 
         // Logarithm base 2. At this point we know that pageSize is a power of two.
-        return (byte) (Integer.SIZE - 1 - Integer.numberOfLeadingZeros(pageSize));
+        return Integer.SIZE - 1 - Integer.numberOfLeadingZeros(pageSize);
     }
 
     private static int validateAndCalculateChunkSize(int pageSize, int maxOrder) {
